+++ scripts/pipeline/pipeline_manager.py
@@ -1,6 +1,5 @@
 import click
-
-from pipeline.pipeline_parameters import PipelineParametersHandler
+from orchestrator import PipelineParametersHandler
 
 
 @click.group(chain=True)
@@ -8,18 +7,16 @@ def pipeline():
     pass
 
 
-def _prepare_handler() -> PipelineParametersHandler:
-    return PipelineParametersHandler.from_env()
-
-
 @pipeline.command("log_params")
 def log_params():
-    _prepare_handler().log_pipeline_params()
+    handler = PipelineParametersHandler.from_env()
+    handler.log_pipeline_params()
 
 
 @pipeline.command("write_dotenv")
 def write_dotenv():
-    _prepare_handler().write_dotenv()
+    handler = PipelineParametersHandler.from_env()
+    handler.write_dotenv()
 
 
 if __name__ == "__main__":
