+++ scripts/tests/pipeline/test_step_gating.py
@@ -32,7 +32,6 @@ def pipeline_env(monkeypatch, tmp_path):
 
 
 def _ctx(**overrides) -> PipelineParametersHandler:
-    os.environ.setdefault("ENV_NAMES", "cluster-01/env-01")
     for key, value in overrides.items():
         os.environ[key] = str(value).lower() if isinstance(value, bool) else str(value)
     return PipelineParametersHandler.from_env()
